move vhdl generating code out of module.py, kept API in Module class, but now the code lives in module_vhdl_gen.py
add curses-menu and PrettyTable to install_requires
